+++ App/Entities/Stop.cs
@@ -6,11 +6,11 @@ using DeLijn.Net.App.Entities.Response;
 namespace DeLijn.Net.App.Entities;
 
 public record Stop(
-    [property: JsonPropertyName("entiteitnummer")][property: JsonConverter(typeof(StringIdConverter))] int EntityId,
-    [property: JsonPropertyName("haltenummer")][property: JsonConverter(typeof(StringIdConverter))] int Id,
+    [property: JsonPropertyName("entiteitnummer")][property: JsonConverter(typeof(StringIdConverter))] short EntityId,
+    [property: JsonPropertyName("haltenummer")][property: JsonConverter(typeof(StringIdConverter))] short Id,
     [property: JsonPropertyName("omschrijving")] string Description,
     [property: JsonPropertyName("omschrijvingLang")] string DescriptionLong,
-    [property: JsonPropertyName("gemeentenummer")][property: JsonConverter(typeof(StringIdConverter))] int MunicipalityId,
+    [property: JsonPropertyName("gemeentenummer")][property: JsonConverter(typeof(StringIdConverter))] short MunicipalityId,
     [property: JsonPropertyName("omschrijvingGemeente")] string MunicipalityDescription,
     //[property: JsonPropertyName("districtCode")] string DistrictCode,
     [property: JsonPropertyName("geoCoordinaat")] GeoCoordinate GeoCoordinate,
